Balaji Physiotherapy Clinic Website
This is  Balaji Physiotherapy Clinic’s official website, which was
built by me. In order to use it, you need to create a local web server,
 (since it is a php file) and then load index.php. Make sure that all
the folders, like images, css, js ,etc are in the same root directory
as index.php. This was my first try at making websites, so hope you
like it!